Improved some details in CellSimilarity concept.
git-svn-id: https://svn.dealii.org/trunk@18467 0785d39b-7218-0410-832d-ea1e28bc413d

+++ deal.II/lac/source/trilinos_sparse_matrix.cc
@@ -406,7 +406,9 @@ namespace TrilinosWrappers
 	  for (int col=0; col < row_length; ++col)
 	    row_indices[col] = sparsity_pattern.column_number (row, col);
 
-	  graph->InsertGlobalIndices (row, row_length, &row_indices[0]);
+	  graph->InsertGlobalIndices (static_cast<int>(row), 
+				      row_length, &row_indices[0]);
+
 	}
 
 				  // Now, fill the graph (sort indices, make
